Move QP comparison to a dedicated module
