+++ apollo-router/src/query_planner/dual_query_planner.rs
@@ -1,0 +1,139 @@
+//! Running two query planner implementations and comparing their results
+
+use std::sync::Arc;
+use std::sync::OnceLock;
+
+use apollo_compiler::ast::Name;
+use apollo_compiler::validation::Valid;
+use apollo_compiler::ExecutableDocument;
+use apollo_compiler::NodeStr;
+use apollo_federation::query_plan::query_planner::QueryPlanner;
+
+use crate::error::format_bridge_errors;
+use crate::executable::USING_CATCH_UNWIND;
+use crate::query_planner::convert::convert_root_query_plan_node;
+use crate::query_planner::render_diff;
+use crate::query_planner::QueryPlanResult;
+
+/// Jobs are dropped if this many are already queued
+const QUEUE_SIZE: usize = 10;
+const WORKER_THREAD_COUNT: usize = 1;
+
+pub(crate) struct BothModeComparisonJob {
+    pub(crate) rust_planner: Arc<QueryPlanner>,
+    pub(crate) document: Arc<Valid<ExecutableDocument>>,
+    pub(crate) operation_name: Option<NodeStr>,
+    pub(crate) js_result: Result<QueryPlanResult, Arc<Vec<router_bridge::planner::PlanError>>>,
+}
+
+type Queue = crossbeam_channel::Sender<BothModeComparisonJob>;
+
+static QUEUE: OnceLock<Queue> = OnceLock::new();
+
+fn queue() -> &'static Queue {
+    QUEUE.get_or_init(|| {
+        let (sender, receiver) = crossbeam_channel::bounded::<BothModeComparisonJob>(QUEUE_SIZE);
+        for _ in 0..WORKER_THREAD_COUNT {
+            let job_receiver = receiver.clone();
+            std::thread::spawn(move || {
+                for job in job_receiver {
+                    job.execute()
+                }
+            });
+        }
+        sender
+    })
+}
+
+impl BothModeComparisonJob {
+    pub(crate) fn schedule(self) {
+        // We use a bounded queue: try_send returns an error when full. This is fine.
+        // We prefer dropping some comparison jobs and only gathering some of the data
+        // rather than consume too much resources.
+        //
+        // Either way we move on and let this thread continue proceed with the query plan from JS.
+        let _ = queue().try_send(self).is_err();
+    }
+
+    fn execute(self) {
+        // TODO: once the Rust query planner does not use `todo!()` anymore,
+        // remove `USING_CATCH_UNWIND` and this use of `catch_unwind`.
+        let rust_result = std::panic::catch_unwind(|| {
+            let name = self.operation_name.clone().map(Name::new).transpose()?;
+            USING_CATCH_UNWIND.set(true);
+            let result = self.rust_planner.build_query_plan(&self.document, name);
+            USING_CATCH_UNWIND.set(false);
+            result
+        })
+        .unwrap_or_else(|panic| {
+            USING_CATCH_UNWIND.set(false);
+            Err(apollo_federation::error::FederationError::internal(
+                format!(
+                    "query planner panicked: {}",
+                    panic
+                        .downcast_ref::<String>()
+                        .map(|s| s.as_str())
+                        .or_else(|| panic.downcast_ref::<&str>().copied())
+                        .unwrap_or_default()
+                ),
+            ))
+        });
+
+        let name = self.operation_name.as_deref();
+        let operation_desc = if let Ok(operation) = self.document.get_operation(name) {
+            if let Some(parsed_name) = &operation.name {
+                format!(" in {} `{parsed_name}`", operation.operation_type)
+            } else {
+                format!(" in anonymous {}", operation.operation_type)
+            }
+        } else {
+            String::new()
+        };
+
+        let is_matched;
+        match (&self.js_result, &rust_result) {
+            (Err(js_errors), Ok(_)) => {
+                tracing::warn!(
+                    "JS query planner error{operation_desc}: {}",
+                    format_bridge_errors(js_errors)
+                );
+                is_matched = false;
+            }
+            (Ok(_), Err(rust_error)) => {
+                tracing::warn!("Rust query planner error{operation_desc}: {}", rust_error);
+                is_matched = false;
+            }
+            (Err(_), Err(_)) => {
+                is_matched = true;
+            }
+
+            (Ok(js_plan), Ok(rust_plan)) => {
+                let js_root_node = js_plan.query_plan.node.as_deref();
+                let rust_root_node = convert_root_query_plan_node(rust_plan);
+                is_matched = js_root_node == rust_root_node.as_ref();
+                if is_matched {
+                    tracing::debug!("JS and Rust query plans match{operation_desc}! 🎉");
+                } else {
+                    tracing::warn!("JS v.s. Rust query plan mismatch{operation_desc}");
+                    if let Some(formatted) = &js_plan.formatted_query_plan {
+                        tracing::debug!(
+                            "Diff of formatted plans:\n{}",
+                            render_diff(&diff::lines(formatted, &rust_plan.to_string()))
+                        );
+                    }
+                    tracing::trace!("JS query plan Debug: {js_root_node:#?}");
+                    tracing::trace!("Rust query plan Debug: {rust_root_node:#?}");
+                }
+            }
+        }
+
+        u64_counter!(
+            "apollo.router.operations.query_planner.both",
+            "Comparing JS v.s. Rust query plans",
+            1,
+            "generation.is_matched" = is_matched,
+            "generation.js_error" = self.js_result.is_err(),
+            "generation.rust_error" = rust_result.is_err()
+        );
+    }
+}
